+++ relay-dynamic-config/src/global.rs
@@ -161,8 +161,7 @@ pub struct Options {
     /// Overall sampling of span extraction.
     ///
     /// This number represents the fraction of transactions for which
-    /// spans are extracted. It applies on top of [`crate::Feature::ExtractCommonSpanMetricsFromEvent`],
-    /// so both feature flag and sample rate need to be enabled to get any spans extracted.
+    /// spans are extracted.
     ///
     /// `None` is the default and interpreted as a value of 1.0 (extract everything).
     ///